requirements.txt
Django
arrow
feedparser
pypandoc
python-slugify
requests